[!!!][TASK] Add charset and collation to all MySQL migrations
This change set adds charset and collation to create table statements
in the existing migrations. This make sure the tables are set up
correctly independent of the database default configuration.

Also migrations generated contain this information since a while,
leading to problems on migration if the database is not using the same
charset and collation.

This is breaking if you have existing tables that do not use the utf8
charset and utf8_unicode_ci collation. To solve this you need to convert
the existing tables. This can be done using the command::

  ./flow database:setcharset

Change-Id: I42ea1b9efe189e102da76bda7f947cce8439d99f
Releases: master
Related: NEOS-800
Reviewed-on: http://review.typo3.org/34590
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Tested-by: Karsten Dambekalns <karsten@typo3.org>
Reviewed-by: Bastian Waidelich <bastian@typo3.org>