[BUGFIX] Link on logo should be to home
The link on the Neos logo is supposed to go to the home page.
In the template a link to the variable ``site`` is created, but
as this variable is never set the link is to the current page.
This change adds the variable to the template, fixing the link.

Change-Id: Iab931a30e769bfdd06fd95dd496785b202794af3
Releases: master, 1.2
Reviewed-on: http://review.typo3.org/36289
Reviewed-by: Bastian Waidelich <bastian@typo3.org>
Reviewed-by: Christian Müller <christian@kitsunet.de>
Tested-by: Christian Müller <christian@kitsunet.de>
Reviewed-by: Aske Ertmann <aertmann@gmail.com>