[TASK] Update exported site
This is an export done on latest master after a clean setup and running
node:repair.

Timestamps are removed since to ensure new dates when imported.

Additionally it's output with the ``tidy`` option to track actual
changes of the export in the future.

Also fixes the issue with unwrapped text paragraphs that trigger a save
on the first click.

Change-Id: I0ffc5fdd47c5d4e3f4c725f8bc7f02d8b91bbc77
Releases: master
Reviewed-on: http://review.typo3.org/37055
Reviewed-by: Aske Ertmann <aertmann@gmail.com>
Tested-by: Aske Ertmann <aertmann@gmail.com>
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Tested-by: Karsten Dambekalns <karsten@typo3.org>