[FEATURE] Frontend Login
This adds a simple *Frontend User Login* form to the site demonstrating
the advanced security features introduced with Neos 2.0.

This adds a dependency to the ``flowpack/neos-frontendlogin`` composer
package that comes with two plugins:

* ``Flowpack.Neos.FrontendLogin:LoginForm`` serving both, a *login*
  form for anonymous users and a *logout* form for authenticated
  Frontend Users
* ``Flowpack.Neos.FrontendLogin:Profile`` posing a very basic way to
  allow users to change their details

Note: There is no technical difference between the "Backend User" that
can login to Neos and a "Frontend User" that is usually a visitor of
your Website. But for the sake of simplicity the FE Login plugin comes
with a dedicated role ``Flowpack.Neos.FrontendLogin:User`` that will
make it easier to distinguish the two cases.

To create a "Frontend User" you can use the provided command:

  ./flow frontenduser:create <username> <password> <first-name> <last-name>

Change-Id: I8bbd30556c79e1cc87db640339c394d67f99cf84
Resolves: NEOS-433
Releases: master
Reviewed-on: http://review.typo3.org/34347
Reviewed-by: Bastian Waidelich <bastian@typo3.org>
Reviewed-by: Andreas Förthner <andreas.foerthner@netlogix.de>
Tested-by: Bastian Waidelich <bastian@typo3.org>
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>