[BUGFIX] Demo site should "correctly" create footer node
The demo site created the "footer" child node on the homepage purely from
the ``Sites.xml``. It was not configured in ``NodeTypes.yaml``.
This is not best practice and could break if we introduce functionality
to remove child nodes that are not configured.
Now the homepage of the demo site is a new NodeType that has ``footer``
configured as childNode.

Change-Id: Ib878d80308e139b4826e291176e4fa59f66c2cfe
Fixes: NEOS-913
Releases: master, 2.0
Reviewed-on: http://review.typo3.org/38802
Reviewed-by: Christian Müller <christian.mueller@typo3.org>
Tested-by: Christian Müller <christian.mueller@typo3.org>
Tested-by: Mr. Hudson <hudson@typo3.org>