[BUGFIX] ContentCollection iterator has a different TypoScript
The change `I06ba604f562685e3a1d5ee844a8adaec7eda20f5` introduced a
backwards incompatible change to the TypoScript paths of
`ContentCollections`. Before it was possible to set the
`iterationName` directly on the `ContentCollection`. With the newly
introduced `TYPO3.Neos:ContentCollectionRenderer` the path for this
moved to `content.iterationName` breaking the image carousel example
on the demo site.

Change-Id: I1ced14b0146c1c0c93e46d166f19683c6e17713e
Releases: master
Reviewed-on: http://review.typo3.org/38937
Reviewed-by: Christian Müller <christian.mueller@typo3.org>
Tested-by: Christian Müller <christian.mueller@typo3.org>
Tested-by: Mr. Hudson <hudson@typo3.org>
Reviewed-by: Aske Ertmann <aertmann@gmail.com>
Tested-by: Aske Ertmann <aertmann@gmail.com>