TASK: Remove fusion-namespaces from Neos.Demo
The namespaces are a source of confusion and using them in
Neos.Demo encourages too many people to use them and run
into confusion.

In addition we are planning to remove the definition of the default
fusion namespace and the transfer of namespaces to includes and
this is part of the preparation for that.