BUGFIX: Fix script tags in document head
self closing script tags are not allowed.
this led to the whole inline editing to fail in the Neos backend.
it is a regression that was introduced with #83.
it took us far too long to find the root cause.
we should stabilze the neos backend for those cases.
wtf.

ps: I also re-added the `defer` attributes that got lost with the
    original change